requirements.txt
FreeSimpleGUI
ffmpeg-python